Fixed missing type info in object register when returning from sys call
When calling a function that returns a reference or handle thru a system
call, the objectType field is not set. So if the function that is called
suspends execution, the state of the VM is incomplete (objectType is
garabe). This leads to a crash if the context is cleaned up, unless
execution is continued.